Move ReadOnlyModBook to moodel package
