+++ src/main/java/seedu/address/model/ModBook.java
@@ -6,7 +6,6 @@ import java.util.List;
 
 import javafx.collections.ObservableList;
 import seedu.address.model.module.Module;
-import seedu.address.model.module.ReadOnlyModBook;
 import seedu.address.model.module.UniqueModuleList;
 
 /**
