+++ src/main/java/seedu/address/model/Model.java
@@ -5,7 +5,6 @@ import java.util.function.Predicate;
 
 import javafx.collections.ObservableList;
 import seedu.address.commons.core.GuiSettings;
-import seedu.address.model.module.ReadOnlyModBook;
 import seedu.address.model.person.Person;
 import seedu.address.model.module.Module;
