+++ src/main/java/seedu/address/model/ReadOnlyModBook.java
@@ -1,4 +1,4 @@
-package seedu.address.model.module;
+package seedu.address.model;
 
 import javafx.collections.ObservableList;
 import seedu.address.model.module.Module;
